move package common under replication
