+++ server/src/main/java/org/opensearch/indices/recovery/PeerRecoveryTargetService.java
@@ -69,10 +69,10 @@ import org.opensearch.index.shard.ShardNotFoundException;
 import org.opensearch.index.store.Store;
 import org.opensearch.index.translog.Translog;
 import org.opensearch.index.translog.TranslogCorruptedException;
-import org.opensearch.indices.common.ShardTargetCollection;
-import org.opensearch.indices.common.ShardTargetCollection.ShardTargetRef;
-import org.opensearch.indices.common.ReplicationLuceneIndex;
-import org.opensearch.indices.common.ReplicationTimer;
+import org.opensearch.indices.replication.common.ShardTargetCollection;
+import org.opensearch.indices.replication.common.ShardTargetCollection.ShardTargetRef;
+import org.opensearch.indices.replication.common.ReplicationLuceneIndex;
+import org.opensearch.indices.replication.common.ReplicationTimer;
 import org.opensearch.tasks.Task;
 import org.opensearch.threadpool.ThreadPool;
 import org.opensearch.transport.ConnectTransportException;
