+++ server/src/main/java/org/opensearch/indices/replication/common/ShardTarget.java
@@ -6,7 +6,7 @@
  * compatible open source license.
  */
 
-package org.opensearch.indices.common;
+package org.opensearch.indices.replication.common;
 
 import org.apache.logging.log4j.Logger;
 import org.opensearch.ExceptionsHelper;
